New post processing method
Tried creating a new method where the model can run it's full prediction and where it comflicted with the mode then use the probablities from the predictions to guess wqhich of the known genres to pick. Didn't work otu as powerfull as i thought. It's not a total loss but needs adjusting.